# pkg/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mezzium API",
    "version": "1.0.0",
    "description": "Proxy JSON-RPC and REST requests to EVM, BTC-like, TRON, Solana and other networks with fail-fast per-node timeouts, health checks, gossip discovery, and admin API for runtime network management."
  },
  "servers": [
    { "url": "https://api.mezzium.net", "description": "api.mezzium.net" },
    { "url": "http://127.0.0.1:8080", "description": "Local default" }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Admin API key for adding networks and nodes at runtime."
      }
    },
    "schemas": {
...
      "Peer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "addr": { "type": "string", "example": "127.0.0.1:8080" }
        }
      },
      "NodeConfig": {
        "type": "object",
        "properties": {
          "url": { "type": "string", "format": "uri" },
          "priority": { "type": "integer", "minimum": 1, "default": 1 },
          "headers": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "example": { "x-api-key": "YOUR_KEY" }
          },
          "tor": { "type": "boolean", "default": false }
        },
        "required": ["url"]
      },
...
      "NetworkConfig": {
        "type": "object",
        "required": ["route", "protocol", "nodes"],
        "properties": {
          "route": { "type": "string", "example": "/matic" },
          "protocol": { "type": "string", "enum": ["evm", "btc", "trx", "sol", "doge", "ltc"] },
          "timeoutMs": { "type": "integer", "example": 1500 },
          "nodes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NodeConfig" }
          }
        }
      },
      "AnnounceRequest": {
        "type": "object",
        "required": ["id", "name", "internal_addr", "timestamp", "signature"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "internal_addr": { "type": "string", "example": "10.0.0.12:8080" },
          "timestamp": { "type": "integer" },
          "signature": { "type": "string", "description": "HMAC-SHA256 hex" }
        }
      },
      "AnnounceResponse": {
        "type": "object",
        "properties": {
          "peers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Peer" }
          }
        }
      },
      "GossipMessage": {
        "type": "object",
        "properties": {
          "from": { "type": "string" },
          "peers": { "type": "array", "items": { "$ref": "#/components/schemas/Peer" } }
        }
      },
      "NodeAdvert": {
        "type": "object",
        "properties": {
          "url": { "type": "string" },
          "priority": { "type": "integer" },
          "protocol": { "type": "string" },
          "alive": { "type": "boolean" },
          "ping": { "type": "integer" }
        }
      },
      "NetworkAdvert": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "protocol": { "type": "string" },
          "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/NodeAdvert" } },
          "ts": { "type": "integer" }
        }
      },
      "StateMessage": {
        "type": "object",
        "properties": {
          "from": { "type": "string" },
          "networks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NetworkAdvert" }
          }
        }
      },
      "Heartbeat": {
        "type": "object",
        "properties": {
          "leader_id": { "type": "string" },
          "timestamp": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/announce": {
      "post": {
        "tags": ["Cluster"],
        "summary": "Announce this node to bootstrap server",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnnounceRequest" },
              "example": {
                "id": "e6a1bdce-7f7e-4aab-9c2b-5b8f2a3e1f00",
                "name": "rpcf-node-1",
                "internal_addr": "10.0.0.12:8080",
                "timestamp": 1732522800,
                "signature": "c0ffee..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of known peers",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnounceResponse" },
                "example": { "peers": [ { "id": "self", "addr": "10.0.0.12:8080" }, { "id": "peer-2", "addr": "10.0.0.13:8080" } ] }
              }
            }
          },
          "403": { "description": "Invalid signature" }
        }
      }
    },
    "/gossip": {
      "post": {
        "tags": ["Cluster"],
        "summary": "Receive peer list via gossip",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GossipMessage" },
              "example": { "from": "node-1", "peers": [ { "id": "node-1", "addr": "10.0.0.12:8080" }, { "id": "node-2", "addr": "10.0.0.13:8080" } ] }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/gossip-state": {
      "post": {
        "tags": ["Cluster"],
        "summary": "Receive best-node state via gossip",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StateMessage" },
              "example": {
                "from": "node-1",
                "networks": [
                  {
                    "name": "ethereum",
                    "protocol": "evm",
                    "nodes": [ { "url": "https://eth.llamarpc.com", "priority": 1, "protocol": "evm", "alive": true, "ping": 123 } ],
                    "ts": 1732522800
                  }
                ]
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/heartbeat": {
      "post": {
        "tags": ["Cluster"],
        "summary": "Record heartbeat from leader node",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Heartbeat" }, "example": { "leader_id": "node-1", "timestamp": 1732522800 } } }
        },
        "responses": { "200": { "description": "Heartbeat recorded" } }
      }
    },
...
    "/admin/networks": {
      "post": {
        "tags": ["Admin"],
        "security": [{ "AdminKey": [] }],
        "summary": "Add a new network",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NetworkConfig" },
              "example": {
                "route": "/matic",
                "protocol": "evm",
                "timeoutMs": 1500,
                "nodes": [
                  { "url": "https://polygon-rpc.com", "priority": 1 },
                  { "url": "https://polygon-mainnet.gateway.tatum.io", "priority": 2, "headers": { "x-api-key": "YOUR_KEY" } }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Network added" },
          "400": { "description": "Invalid payload or no healthy nodes" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/admin/networks/bulk": {
      "post": {
        "tags": ["Admin"],
        "security": [{ "AdminKey": [] }],
        "summary": "Add multiple networks in bulk",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/NetworkConfig" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk add result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "route": { "type": "string" },
                      "status": { "type": "string" },
                      "reason": { "type": "string" },
                      "healthyNodes": { "type": "integer" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/proxy/ethereum/estimateGas": {
      "post": {
        "tags": ["Public"],
        "summary": "Estimate gas for a transaction via eth_estimateGas",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "type": "object" }
              },
              "example": [
                {
                  "from": "0x0000000000000000000000000000000000000000",
                  "to": "0x0000000000000000000000000000000000000000",
                  "value": "0x1",
                  "data": "0x"
                }
              ]
            }
          }
        },
        "responses": {
          "200": { "description": "Estimated gas" }
        }
      }
    },
...
  }
}